+++ libs/server/common/src/activity/activity.dto.ts
@@ -1,0 +1,3 @@
+import { UserActivity } from './activity.model';
+
+export type CreateActivityInput = Omit<UserActivity, 'id' | 'createdAt' | 'updatedAt'>;
