+++ libs/server/common/src/activity/activity.util.ts
@@ -1,8 +1,8 @@
-import { PullRequestEventData, PushEventData, IssueEventData } from '../models';
-import { WebhookEvent } from '../models/event.model';
-import { ActivityData, ActivityType } from '../models/user.model';
+import { IssueEventData, PullRequestEventData, PushEventData } from "../core";
+import { WebhookEvent } from "../event/event.model";
+import { ActivityData, ActivityType } from "./activity.model";
 
-export class EventUtils {
+export class ActivityUtils {
   static mapToActivityType(event: WebhookEvent): ActivityType {
     const eventType = event.eventType;
     const eventAction = (event.data as any)?.action;
@@ -65,33 +65,6 @@ export class EventUtils {
       }
       default:
         return undefined;
-    }
-  }
-
-  static generateUserFacingDescription(event: WebhookEvent): string {
-    const eventType = event.eventType;
-    const eventData = event.data as PushEventData | PullRequestEventData | IssueEventData;
-    const repoName = eventData.repository.name;
-
-    switch (eventType) {
-      case 'push': {
-        const data = eventData as PushEventData;
-        const branch = data.branch.replace('refs/heads/', '');
-        return `Committed ${data.commitCount} time(s) to ${repoName} (${branch}) (GitHub)`;
-      }
-      case 'pull_request': {
-        const data = eventData as PullRequestEventData;
-        const action = data.action === 'closed' && data.state === 'closed' ? 'merged' : data.action;
-        return `${action.charAt(0).toUpperCase() + action.slice(1)} PR #${data.prNumber} in ${repoName} (GitHub)`;
-      }
-      case 'issue': {
-        const data = eventData as IssueEventData;
-        return `${data.action.charAt(0).toUpperCase() + data.action.slice(1)} issue #${
-          data.issueNumber
-        } in ${repoName} (GitHub)`;
-      }
-      default:
-        return 'Performed an action on GitHub';
     }
   }
 }
