+++ libs/server/common/src/core/interfaces/index.ts
