+++ libs/server/common/src/core/models/index.ts
@@ -1,0 +1,4 @@
+export * from './common.model';
+export * from './gamification-domain.model';
+export * from './github-shared.model';
+
