+++ libs/server/common/src/core/services/index.ts
@@ -1,7 +1,4 @@
-export * from './activity.service';
 export * from './base.service';
 export * from './counter.service';
 export * from './database.service';
-export * from './event.service';
 export * from './storage.service';
-export * from './user.service';
