+++ libs/server/common/src/event/event.dto.ts
@@ -1,0 +1,3 @@
+import { WebhookEvent } from "./event.model";
+
+export type CreateEventInput = Omit<WebhookEvent, 'id' | 'createdAt' | 'updatedAt'>;
