+++ libs/server/common/src/firebase/logger.util.ts
