--- libs/server/common/src/models/index.ts
@@ -1,5 +1,0 @@
-export * from './common.model';
-export * from './event.model';
-export * from './github-shared.model';
-export * from './user.model';
-export * from './gamification-domain.model';
