+++ libs/server/common/src/user/user.dto.ts
@@ -1,0 +1,6 @@
+export interface CreateUserInput {
+  uid?: string;
+  email: string | null;
+  displayName: string | null;
+  photoUrl: string | null;
+}
