+++ libs/server/common/src/user/user.model.ts
@@ -1,0 +1,10 @@
+import { BaseDocument } from "../core/models/common.model";
+
+export interface User extends BaseDocument {
+  email: string;
+  displayName: string;
+  photoUrl: string;
+  lastLogin: string;
+  active: boolean;
+  uid?: string;
+}
