--- libs/server/common/src/utils/firebase/index.ts
@@ -1,1 +1,0 @@
-export * from './firestore.util';
