--- libs/server/common/src/utils/index.ts
@@ -1,4 +1,0 @@
-export * from './converters.util';
-export * from './firebase';
-export * from './logger.util';
-export * from './time';
