--- libs/server/common/src/utils/time/index.ts
@@ -1,1 +1,0 @@
-export * from './time.utils';
